+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Next 0.7.0-next.2
 ## Features:
-- [BDN-626](https://appodeal.atlassian.net/browse/BDN-626) Integrated MyTarget as Ad Adapter
+- [BDN-626](https://appodeal.atlassian.net/browse/BDN-626) Integrated VkAds (myTarget) as Ad Adapter
 
 # Next 0.7.0-next.1
 ## Features: